4-21

+++ osa4/blogilista/routes/blogs.js
@@ -48,6 +48,23 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    try {
+        // use var as a band-aid to make decodedToken visible outside try block because const doesnt seem to work
+        var decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!request.token || !decodedToken) {
+            return response.status(401).json({ error: "token missing or invalid" })
+        }
+    } catch (error) {
+        return response.status(401).json({ error: "token invalid" })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        return response.status(404).json({ error: "blog not found" })
+    }
+    if (blog.user.toString() !== decodedToken.id) {
+        return response.status(403).json({ error: "only owner of blog can delete it" })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
